Start of new version of fnm
this version hopes to allow the combined use of fnm elements and abaqus
default capabilities. so the preproc will be updated to allow a single
fnm part, together with other parts made of abaqus default elements.
All nsets of the fnm part are defined in the part (NOT in the
assembly). Only the predelam elset (introduced from ver-12-2015) is
allowed in the fnm part. In this way, no modification is needed for the
assembly and bcd sections of the input file.

Another capability which this version aims to bring, is the modelling
of compressive failure mechanisms. Slanted matrix crack has already
been implemented in the prev. version (ver-12-2015). Friction could be
included in the matrix compressive failure criterion. More innovations
could be done for fibre compressive failure (kink-band formation).